+++ changelog.md
@@ -5,6 +5,7 @@
 - Fixed null deref due to Plot Armor checking whether a non-existent room is a boss foyer
 - Fixed null deref in Ki Blast when trying to apply knockback to enemies without KnockbackDoers
 - Fixed null deref in Ki Blast when used against the Dragun
+- Fixed not being able to play the Glockarina when using a controller
 
 ## 1.11.4 (2024-04-12)